Påbörjat konstant navbar, städat upp kod
