+++ src/components/layout/Footer.tsx
@@ -1,6 +1,6 @@
 const Footer = () => {
     return (
-    <footer className="footer footer-center p-3">
+    <footer className="footer footer-center p-3 bg-base-100">
     <aside>
         <p className="font-bold opacity-30">
         ARKINO - Social Entertainment Systems
