+++ src/pages/_app.tsx
@@ -3,9 +3,9 @@ import { SessionProvider } from "next-auth/react";
 import { type AppType } from "next/app";
 import { useEffect } from 'react';
 import { api } from "~/utils/api";
-import Navbar from "~/components/NavBar";
-import { AuthContext } from "~/components/authContext";
-import Footer from "~/components/Footer";
+import Navbar from "~/components/layout/NavBar";
+import { AuthContext } from "~/components/auth/authContext";
+import Footer from "~/components/layout/Footer";
 
 
 import "~/styles/globals.css";
@@ -16,15 +16,15 @@ const MyApp: AppType<{ session: Session | null }> = ({
 }) => {
 
   useEffect(() => {
-    document.documentElement.setAttribute('data-theme', 'night');
+    document.documentElement.setAttribute('data-theme', 'light');
   }, []);
 
   return (
     <SessionProvider session={session}>
       <AuthContext.Provider value={{ session }}>
-        <div className="flex flex-col min-h-screen bg-base-200">
+        <div className="flex flex-col h-screen max-h-screen bg-base-200">
           <Navbar />
-          <div className="flex flex-grow p-5"> 
+          <div className="flex flex-auto"> 
             <Component {...pageProps} />
           </div>
           <Footer />
