+++ tailwind.config.ts
@@ -10,6 +10,6 @@ export default {
   },
   plugins: [require('@tailwindcss/typography'),require("daisyui")],
   daisyui: {
-    themes: ["light", "dark", "night", "cupcake", "bumblebee", "emerald", "corporate", "synthwave", "retro", "cyberpunk", "valentine", "halloween", "garden", "forest", "aqua"], 
+    themes: ["light", "dark", "night", "cupcake", "bumblebee", "emerald", "corporate", "synthwave", "nord", "emerald"], 
   },
 }
